Added comments to make it easier to understand
The process of adding comments to code that I read helps me make sense of it even if I never read them again.  Comments that are just a question mark are place holders to return to when I have a better idea of what that piece of code is for.