Now using wasabeef rich editor in master.  Seems to work so far.  Need to get rid of the Sticky note stuff as it no longer makes sense.
Added flag for AutomaticMerge.  This should probably be an enum together with the stickynotes because they are mutually exclusive.

Added Db and associated classes as a container for the various classes representing tables because we now have two more tables to deal with to support the vector clocks.  Moved NotesDbHelper to the new Db class.

Added MakeMessageWithAttachment function to begin storing messages in attachments instead of inline this should simplify the business of handling composite documents.

Added @NonNull to some argument lists.